Changed code
Type of variables from Float to Double
Added init commentary
Changed init
